Moves render_movie_in_nuke into a hook.
