+++ app.py
@@ -1,11 +1,11 @@
 # Copyright (c) 2013 Shotgun Software Inc.
-# 
+#
 # CONFIDENTIAL AND PROPRIETARY
-# 
-# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit 
+#
+# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit
 # Source Code License included in this distribution package. See LICENSE.
-# By accessing, using, copying or modifying this work you indicate your 
-# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights 
+# By accessing, using, copying or modifying this work you indicate your
+# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
 """
@@ -16,13 +16,13 @@ import copy
 import os
 import sys
 
-import nuke
 import tank
 import tank.templatekey
 from tank.platform.qt import QtCore
 
 
 class QuicktimeGenerator(tank.platform.Application):
+
     def __init__(self, *args, **kwargs):
         super(QuicktimeGenerator, self).__init__(*args, **kwargs)
         self._logo = None
@@ -30,11 +30,11 @@ class QuicktimeGenerator(tank.platform.Application):
         self._font = None
 
     def init_app(self):
-        
+
         self._burnin_nk = os.path.join(self.disk_location, "resources", "burnin.nk")
         self._font = os.path.join(self.disk_location, "resources", "liberationsans_regular.ttf")
 
-        # If the slate_logo supplied was an empty string, the result of getting 
+        # If the slate_logo supplied was an empty string, the result of getting
         # the setting will be the config folder which is invalid so catch that
         # and make our logo path an empty string which Nuke won't have issues with.
         if os.path.isfile( self.get_setting("slate_logo", "") ):
@@ -47,27 +47,27 @@ class QuicktimeGenerator(tank.platform.Application):
         if sys.platform == "win32":
             self._font = self._font.replace(os.sep, "/")
             self._logo = self._logo.replace(os.sep, "/")
-            self._burnin_nk = self._burnin_nk.replace(os.sep, "/")        
+            self._burnin_nk = self._burnin_nk.replace(os.sep, "/")
 
     def render_and_submit(self, template, fields, first_frame, last_frame, sg_publishes, sg_task,
-                          comment, thumbnail_path, progress_cb):
+                          comment, thumbnail_path, progress_cb, **kwargs):
         """
         Main application entry point to be called by other applications / hooks.
 
         :template: SGTK Template object. The template defining the path where
                    frames should be found.
-                        
+
         :fields: Fields to be used to fill out the template with.
-               
+
         :first_frame: int. The first frame of the sequence of frames.
-                        
+
         :last_frame: int. The last frame of the sequence of frames.
-                        
+
         :sg_publishes: A list of shotgun published file objects to link the publish against.
-                     
+
         :sg_task: A Shotgun task object to link against. Can be None.
-                        
-        :comment: str. A description to add to the Version in Shotgun. 
+
+        :comment: str. A description to add to the Version in Shotgun.
 
         Returns the Version that was created in Shotgun.
         """
@@ -103,16 +103,28 @@ class QuicktimeGenerator(tank.platform.Application):
 
         # Render and Submit
         progress_cb(20, "Rendering movie")
-        self._render_movie_in_nuke(fields, path, output_path, width, height, first_frame, last_frame)
+        self.app.execute_hook("hook_render_movie_in_nuke",
+                              burnin_nk=self._burnin_nk,
+                              font=self._font,
+                              logo=self._logo,
+                              context=self.context
+                              fields=fields,
+                              path=path,
+                              output_path=output_path,
+                              width=width,
+                              height=height,
+                              first_frame=first_frame,
+                              last_frame=last_frame,
+                              **kwargs)
 
         progress_cb(50, "Creating Shotgun Version")
-        sg_version = self._submit_version(path, 
+        sg_version = self._submit_version(path,
                                           output_path,
-                                          sg_publishes, 
-                                          sg_task, 
-                                          comment, 
+                                          sg_publishes,
+                                          sg_task,
+                                          comment,
                                           store_on_disk,
-                                          first_frame, 
+                                          first_frame,
                                           last_frame)
 
         # Upload in a new thread and make our own event loop to wait for the
@@ -123,11 +135,11 @@ class QuicktimeGenerator(tank.platform.Application):
         thread.finished.connect(event_loop.quit)
         thread.start()
         event_loop.exec_()
-        
+
         # log any errors generated in the thread
         for e in thread.get_errors():
             self.log_error(e)
-            
+
         # Remove from filesystem if required
         if not store_on_disk and os.path.exists(output_path):
             os.unlink(output_path)
@@ -139,10 +151,10 @@ class QuicktimeGenerator(tank.platform.Application):
         """
         Create a version in Shotgun for this path and linked to this publish.
         """
-        
+
         # get current shotgun user
         current_user = tank.util.get_current_user(self.tank)
-        
+
         # create a name for the version based on the file name
         # grab the file name, strip off extension
         name = os.path.splitext(os.path.basename(path_to_movie))[0]
@@ -150,7 +162,7 @@ class QuicktimeGenerator(tank.platform.Application):
         name = name.replace("_", " ")
         # and capitalize
         name = name.capitalize()
-        
+
         # Create the version in Shotgun
         data = {
             "code": name,
@@ -184,134 +196,6 @@ class QuicktimeGenerator(tank.platform.Application):
         sg_version = self.tank.shotgun.create("Version", data)
         self.log_debug("Created version in shotgun: %s" % str(data))
         return sg_version
-
-    def _render_movie_in_nuke(self, fields, path, output_path, width, height, first_frame, last_frame):
-        """
-        Use Nuke to render a movie. This assumes we're running _inside_ Nuke.
-        """
-        output_node = None
-
-        # create group where everything happens
-        group = nuke.nodes.Group()
-        
-        # now operate inside this group
-        group.begin()
-        try:
-            # create read node
-            read = nuke.nodes.Read(name="source", file=path)
-            read["on_error"].setValue("black")
-            read["first"].setValue(first_frame)
-            read["last"].setValue(last_frame)
-            
-            # now create the slate/burnin node
-            burn = nuke.nodePaste(self._burnin_nk) 
-            burn.setInput(0, read)
-        
-            # set the fonts for all text fields
-            burn.node("top_left_text")["font"].setValue(self._font)
-            burn.node("top_right_text")["font"].setValue(self._font)
-            burn.node("bottom_left_text")["font"].setValue(self._font)
-            burn.node("framecounter")["font"].setValue(self._font)
-            burn.node("slate_info")["font"].setValue(self._font)
-        
-            # add the logo
-            burn.node("logo")["file"].setValue(self._logo)
-            
-            # format the burnins
-            version_padding_format = "%%0%dd" % self.get_setting("version_number_padding")
-            version_str = version_padding_format % fields.get("version", 0)
-            
-            
-            if self.context.task:
-                version = "%s, v%s" % (self.context.task["name"], version_str)
-            elif self.context.step:
-                version = "%s, v%s" % (self.context.step["name"], version_str)
-            else:
-                version = "v%s" % version_str
-            
-            burn.node("top_left_text")["message"].setValue(self.context.project["name"])
-            burn.node("top_right_text")["message"].setValue(self.context.entity["name"])
-            burn.node("bottom_left_text")["message"].setValue(version)
-            
-            # and the slate            
-            slate_str =  "Project: %s\n" % self.context.project["name"]
-            slate_str += "%s: %s\n" % (self.context.entity["type"], self.context.entity["name"])
-            slate_str += "Name: %s\n" % fields.get("name", "Unnamed").capitalize()
-            slate_str += "Version: %s\n" % version_str
-            
-            if self.context.task:
-                slate_str += "Task: %s\n" % self.context.task["name"]
-            elif self.context.step:
-                slate_str += "Step: %s\n" % self.context.step["name"]
-            
-            slate_str += "Frames: %s - %s\n" % (first_frame, last_frame)
-            
-            burn.node("slate_info")["message"].setValue(slate_str)
-
-            # create a scale node
-            scale = self._create_scale_node(width, height)
-            scale.setInput(0, burn)                
-
-            # Create the output node
-            output_node = self._create_output_node(output_path)
-            output_node.setInput(0, scale)
-        finally:
-            group.end()
-    
-        if output_node:
-            # Make sure the output folder exists
-            output_folder = os.path.dirname(output_path)
-            self.ensure_folder_exists(output_folder)
-            
-            # Render the outputs, first view only
-            nuke.executeMultiple([output_node], ([first_frame-1, last_frame, 1],), [nuke.views()[0]])
-
-        # Cleanup after ourselves
-        nuke.delete(group)
-    
-    @staticmethod
-    def _create_scale_node(width, height):
-        """
-        Create the Nuke scale node to resize the content.
-        """
-        scale = nuke.nodes.Reformat()
-        scale["type"].setValue("to box")
-        scale["box_width"].setValue(width)
-        scale["box_height"].setValue(height)
-        scale["resize"].setValue("fit")
-        scale["box_fixed"].setValue(True)
-        scale["center"].setValue(True)
-        scale["black_outside"].setValue(True)
-        return scale
-
-    @staticmethod
-    def _create_output_node(path):
-        """
-        Create the Nuke output node for the movie.
-        """
-        node = nuke.nodes.Write()
-
-        # Example output settings
-        #
-        # These are either hard coded by a studio in a quicktime generation app
-        # itself (like here) or part of the configuration - however there is
-        # often the need to have special rules to for example handle multi
-        # platform cases.
-        #
-
-        if sys.platform in ["darwin", "win32"]:
-            # On the mac and windows, we use the quicktime codec
-            node["file_type"].setValue("mov")
-            node["codec"].setValue("jpeg")
-
-        elif sys.platform == "linux2":
-            # On linux, use ffmpeg
-            node["file_type"].setValue("ffmpeg")
-            node["format"].setValue("MOV format (mov)")
-
-        node["file"].setValue(path.replace(os.sep, "/"))
-        return node        
-
 
 class UploaderThread(QtCore.QThread):
     """
