+++ info.yml
@@ -1,17 +1,25 @@
 # Copyright (c) 2013 Shotgun Software Inc.
-# 
+#
 # CONFIDENTIAL AND PROPRIETARY
-# 
-# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit 
+#
+# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit
 # Source Code License included in this distribution package. See LICENSE.
-# By accessing, using, copying or modifying this work you indicate your 
-# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights 
+# By accessing, using, copying or modifying this work you indicate your
+# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
 # Metadata defining the behaviour and requirements for this app
 
 # expected fields in the configuration file for this app
 configuration:
+
+    hook_render_movie_in_nuke:
+        type: hook
+        default_value: "render_movie_in_nuke"
+        parameters: [burnin_nk, font, logo, context, fields, path, output_path, width, height, first_frame, last_frame]
+        description: Hook used to perform the actual nuke rendering.  This hook can be
+                     overridden to use your own custom logic.
+
     upload_to_shotgun:
         type: bool
         default_value: true
@@ -67,7 +75,7 @@ configuration:
     slate_logo:
         type: config_path
         description: This is the path to an image to use on the slate such as a
-                     company logo. The supplied image will be reformated into a 
+                     company logo. The supplied image will be reformated into a
                      400 pixel box and the lower left corner of the logo will be
                      aligned 100 pixels to the right and 100 pixels above the
                      lower left corner of the slate. You can use an image with
@@ -79,14 +87,14 @@ configuration:
 # the Shotgun fields that this app needs in order to operate correctly
 requires_shotgun_fields:
 
-# More verbose description of this item 
+# More verbose description of this item
 display_name: "Review Submission"
 description: "Provides functionality to submit image sequences to Shotgun for review."
 
 # Required minimum versions for this item to run
 requires_shotgun_version:
 requires_core_version: "v0.14.0"
-requires_engine_version: 
+requires_engine_version:
 
 # For now this app is only compatible with Nuke.
 supported_engines: ["tk-nuke"]
